Fixed WebAssembly Mime-Type Mapping in Spring
When serving the app via Spring, the app would not work in Firefox.
After intensive debugging, I realised the problem was with instantiating the WebAssembly module with the Firefox engine.
The problem was only with Spring, not reproducible with the Webpack Dev Server.
The problem was, that Spring returned the wrong Mime type for WebAssembly. At the moment Spring does not support serving WebAssembly if you do not add a Mime Mapping for WebAssembly. I added one, but a wrong one. Instead of "application/wasm; charset=utf-8", which Chromium supports, the correct one is "application/wasm". This fixed the problem.